+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - remove dependencies that are now not needed anymore (`punycode`, `simple-markdown`)
 - dont load quoted messages asynchronously
 - update deltachat-node to v1.65.0
+- Fetch more messages if as close as 200px to top of MessageList
 - Join group via qr-code is now async (group already opens, no wait time)
 
 ### Fixed