+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 - Fix messagelist not being at the correct position after fetching more messages
 - Fix group names/contact names & contact email addresses not being selectable
 - Fix scanning QRCode again while the Dialog is still open
+- Fix react warning in about dialog (#2428)
 
 ### Changed
 - use strict typescript for ui code